Fetch resolves with the resource [resolves #30]
EmberResource#fetch now returns a promise that resolves with the
API data (as before) and the resource itself. This makes it easy to
pass the promise itself around your application and still have
access to the resource once it finishes fetching.

I also documented the signatures of #fetch, #save, and #destroyResource.