Correct scope in ResourceCollection#parse
`ResourceCollection#parse` iterated over each item in the JSON array
and called `this.type.parse` on it. Unfortunately, `this.type.parse` was
the result of an `Ember.wrap` call, which sets `this._super` for
inheritance support. Because we weren't passing a `this` pointer to
`map`, it was being called on `window`, which caused the wrapped
`_super` call to leak.