destroyResource defers instance destruction
When destroying a resource via AJAX, applications often need to use the
Em.Resource object in the UI for a little while after the AJAX call
completes.

This patch lets the object linger along for the run loop in which
the AJAX call completes, so it's not weird to use it for rendering some
UI, or running some application logic.

The object is destroyed in the next run-loop, which will be the new
contract for destroyResource.